requirements.txt
qrcode
pillow
pyzbar
